+++ src/libtcod_c.c
@@ -40,7 +40,6 @@
  *    You can get the development libraries from:
  *    https://www.libsdl.org/download-2.0.php
  *  * Add `GL` to be dynamically linked.
- *  * Compile and link libtcod's src/vendor/glad.c file.
  *  * Compile and link libtcod's src/vendor/lodepng.cpp file.
  *  * Compile and link `src/vendor/utf8proc/utf8proc.c`.
  *  * Link with `zlib`, and add its include directory.  You can include